Separate interface from class, error logging
